requirements.txt
altair==6.0.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dacite==1.9.2
filetype==1.2.0
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
ImageHash==4.3.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.0
minify_html==0.18.1
multimethod==1.12
narwhals==2.14.0
networkx==3.6.1
numba==0.62.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
patsy==1.0.2
phik==0.12.5
pillow==12.0.0
protobuf==6.33.2
puremagic==1.30
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
statsmodels==0.14.6
streamlit==1.52.2
streamlit-ydata-profiling==0.2.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
visions==0.8.1
watchdog==3.0.0
wordcloud==1.9.4
ydata-profiling==4.18.0